Volta o block a links com "d"

+++ src/eventos/messageCreate.js
@@ -48,13 +48,14 @@ module.exports = async (client, message) => {
     };
 
     //Bloqueia Links
-    for (let links of ['discord.gg/', 'discord.com/invite/', 'bit.ly/'])
+    for (let links of ['discord.gg/', 'discord.com/invite/', 'bit.ly/', 'Discord Nitro for Free', 'nitro for 3 months', 'scord.com', 'https://d'])
         if (message.content.toLowerCase().includes(links)) {
             if (message.author.bot) return;
             if (message.member.permissions.has('KICK_MEMBERS')) return false;
             if (message.content.includes('discord.com/channels')) return false;
             if (message.content.includes('discord.gg/rederevo')) return false;
             if (message.content.includes('discord.com')) return false;
+            if (message.content.includes('docs.google')) return false;
             if (message.content.toLowerCase().includes(`@everyone`)) return false;
             if (message.channel.id === "845501522166153226") return false;
             await message.delete()
